pyproject.toml
[project]
name = "leetcode-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.14",
    "beautifulsoup4>=4.13.4",
    "dotenv>=0.9.9",
    "google-genai>=1.26.0",
    "google-generativeai>=0.8.5",
    "numpy>=2.3.1",
    "ollama>=0.5.1",
    "psycopg2>=2.9.10",
    "psycopg2-binary>=2.9.10",
    "pytest-playwright>=0.7.0",
    "requests>=2.32.4",
    "sentence-transformers>=5.0.0",
    "streamlit>=1.47.1",
    "transformers>=4.53.2",
]
